Removed some signed/unsigned comparisons causing Warning when compiling on Windows

+++ src/dataLoader.cpp
@@ -11,7 +11,7 @@
 void dataLoader::convertSEXPToMap(SEXP RData, std::map<std::string, std::string>* ptrMap) {
 	Rcpp::CharacterVector data(RData);
 
-	for (size_t i = 0; i < data.size(); i = i + 2) {
+	for (int i = 0; i < data.size(); i = i + 2) {
 		std::string key(data[i]);
 		std::string value(data[i+1]);
 		(*ptrMap)[key] = value;
@@ -21,7 +21,7 @@ void dataLoader::convertSEXPToMap(SEXP RData, std::map<std::string, std::string>
 void dataLoader::convertSEXPToVector(SEXP RData, std::vector<std::string>* ptrVector) {
 	Rcpp::CharacterVector data(RData);
 
-	for (size_t i = 0; i < data.size(); i++) {
+	for (int i = 0; i < data.size(); i++) {
 		std::string value(data[i]);
 		ptrVector->push_back(value);
 	}
@@ -30,7 +30,7 @@ void dataLoader::convertSEXPToVector(SEXP RData, std::vector<std::string>* ptrVe
 void dataLoader::convertSEXPToDeque(SEXP RData, std::deque<std::string>* ptrDeque) {
 	Rcpp::CharacterVector data(RData);
 
-	for (size_t i = 0; i < data.size(); i = i + 2) {
+	for (int i = 0; i < data.size(); i = i + 2) {
 		std::string value(data[i]);
 		ptrDeque->push_back(value);
 	}
